Move Variant and Loader to API
