+++ src/main/java/com/gtnewhorizons/angelica/api/Loader.java
@@ -1,12 +1,12 @@
-package com.gtnewhorizons.angelica.models.json;
+package com.gtnewhorizons.angelica.api;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.gtnewhorizons.angelica.loading.AngelicaTweaker;
+import com.gtnewhorizons.angelica.models.json.JsonModel;
 import com.gtnewhorizons.angelica.utils.Callback;
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
-import net.minecraft.block.Block;
 import net.minecraft.client.Minecraft;
 import net.minecraft.util.ResourceLocation;
 
@@ -18,19 +18,9 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
-/**
- * Model loading should proceed as follows: <ul>
- * <li>Models to load are registered in PREINIT.</li>
- * <li>All registered models are recursively loaded in POSTINIT.</li>
- * <li>Models registered for baking are baked on the first tick of the game.</li>
- *</ul>
- * <p>Alternatively, you can register blocks with blockstate variants in preinit, and the rest is handled by
- * {@link Loader}.
- * <p>As for icons, register them in {@link Block#registerBlockIcons}. Whatever gets them in the block texture atlas.
- */
 public class Loader {
 
-    static final Gson GSON = new GsonBuilder()
+    private static final Gson GSON = new GsonBuilder()
         .registerTypeAdapter(JsonModel.class, new JsonModel.Deserializer()).create();
     private static final List<ResourceLocation> unloadedModels = new ObjectArrayList<>();
     private static final Map<ResourceLocation, JsonModel> loadedModels = new Object2ObjectOpenHashMap<>();
@@ -113,7 +103,7 @@ public class Loader {
             c.run();
     }
 
-    public static JsonModel getModel(Variant loc) {
+    public static QuadProvider getModel(Variant loc) {
         return modelsToBake.get(loc);
     }
 }
