+++ src/main/java/com/gtnewhorizons/angelica/models/template/Model4Rot.java
@@ -1,8 +1,8 @@
 package com.gtnewhorizons.angelica.models.template;
 
-import com.gtnewhorizons.angelica.models.json.JsonModel;
-import com.gtnewhorizons.angelica.models.json.Loader;
-import com.gtnewhorizons.angelica.models.json.Variant;
+import com.gtnewhorizons.angelica.api.QuadProvider;
+import com.gtnewhorizons.angelica.api.Loader;
+import com.gtnewhorizons.angelica.api.Variant;
 import net.minecraft.util.ResourceLocation;
 
 /**
@@ -10,7 +10,7 @@ import net.minecraft.util.ResourceLocation;
  */
 public class Model4Rot {
 
-    public final JsonModel[] models = new JsonModel[4];
+    public final QuadProvider[] models = new QuadProvider[4];
     private final Variant[] modelIds;
 
     public Model4Rot(ResourceLocation modelLoc) {
