+++ src/main/java/com/gtnewhorizons/angelica/proxy/ClientProxy.java
@@ -9,7 +9,7 @@ import com.gtnewhorizons.angelica.glsm.GLStateManager;
 import com.gtnewhorizons.angelica.glsm.debug.OpenGLDebugging;
 import com.gtnewhorizons.angelica.hudcaching.HUDCaching;
 import com.gtnewhorizons.angelica.models.VanillaModels;
-import com.gtnewhorizons.angelica.models.json.Loader;
+import com.gtnewhorizons.angelica.api.Loader;
 import com.gtnewhorizons.angelica.render.CloudRenderer;
 import com.gtnewhorizons.angelica.rendering.AngelicaBlockSafetyRegistry;
 import com.gtnewhorizons.angelica.utils.AssetLoader;
